% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles2D.R
\docType{data}
\name{angles2D}
\alias{angles2D}
\title{Standard Angles 2D Data Frame}
\format{
An object of class \code{data.frame} with 16 rows and 4 columns.
}
\usage{
angles2D
}
\description{
A test unit data frame consisting of 42 rows and 4 columns with standard angles in Cartesian coordinates as item loadings (columns denoted \code{a1} and \code{a2}) oriented in both positive and negative directions in a two-dimensional space.
The distance from the origin is set by \verb{d = 0,5} (3rd column) on all rows, which refers to the parameter related to difficulty in the compensatory model.
The last two columns contain the angles converted to polar coordinates with \code{Theta} representing the polar angle.
Running the data frame in \code{D2mirt()} converts the angles into polar coordinates and can be used to check functionality in the package.
}
\examples{
data(angles2D)
}
\keyword{datasets}
