% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D2mirt.R
\name{D2mirt}
\alias{D2mirt}
\title{2D DMIRT Model Estimation}
\usage{
D2mirt(
  x,
  modid = NULL,
  model = NULL,
  con.items = NULL,
  con.pol = NULL,
  cap = 1e-04,
  itemtype = "graded",
  method.mirt = "QMCEM",
  method.fscores = "EAP",
  QMC = TRUE,
  ...
)
}
\arguments{
\item{x}{A data frame with items in rows and model parameters in columns containing raw response data as integer values or factor loadings.
Input can also be an S4 object of class 'SingleGroupClass' exported from \link[mirt:mirt]{mirt::mirt} (Chalmers, 2012).
The data frame with factor loadings must have two columns with loading parameters and at least one column for difficulty parameters.}

\item{modid}{Use either the two model identification items from \link{modid} as a combined vector or use nested list of item indicators to fit an orthogonal model (see examples below).
The default is \code{modid = NULL}.}

\item{model}{The user has the option of imputing a model specification schema used in the call to \link[mirt:mirt]{mirt::mirt} (Chalmers, 2012).
The default is \code{model = NULL}.}

\item{con.items}{Optional. Nested lists with integer values as item indicators to identify constructs. The default is \code{con.items = NULL}.}

\item{con.pol}{Optional. Polar angles to identify constructs indicated using integers (e.g., 45 for a single 45 degrees construct and c(30, 45, 60) for three construct angles in polar coordinate form). The default is \code{con.pol = NULL}.}

\item{cap}{Lower bound for standard error to avoid dividing by infinity. The default is \code{cap = 0.0001}.}

\item{itemtype}{What item type to use in the function call. Available options are \code{"2PL"} and \code{"graded"}. The default is \code{itemtype = "graded"}.}

\item{method.mirt}{Estimation algorithm for \link[mirt:mirt]{mirt::mirt} (Chalmers, 2012) to fit the model. The default is \code{method.mirt = "QMCEM"}.}

\item{method.fscores}{Factor estimation algorithm for \link[mirt:fscores]{mirt::fscores} (Chalmers, 2012) for extracting respondent trait scores. The default is \code{method.fscores = "EAP"}.}

\item{QMC}{Integration method for \link[mirt:fscores]{mirt::fscores} (Chalmers, 2012). The default is \code{QMC = TRUE}.}

\item{...}{Any additional arguments passed to mirt().}
}
\value{
A S3 object of class \code{D2mirt} with lists of \emph{a} and \emph{d} parameters from the M2PL or MGRM estimation, multidimensional difficulty (MDIFF), multidimensional discrimination (MDISC), direction cosines and degrees for vector angles, construct lists, vector coordinates, and respondent trait scores.
}
\description{
Descriptive multidimensional item response theory model estimation (DMIRT; Reckase, 2009, 1985, Reckase and McKinley, 1991) for dichotomous and polytomous items restricted to two dimensions.
}
\details{
The \code{D2mirt()} is a two-dimensional version of the \code{D3mirt()} function and takes in model parameters from a compensatory two-dimensional multidimensional two-parameter logistic model (M2PL) or a multidimensional graded
response model (MGRM). See \link{D3mirt} or package vignette for more details on function calls and options.
}
\examples{
\donttest{
# Load data
data("anes0809offwaves")
x <- anes0809offwaves
x <- x[, 3:22] # Remove columns for age and gender

# Call to modid() using two dimensions
a <- modid(x, factors = 2)

# Show summary of results
summary(a)

# Call to D2mirt() using W7Q10 as model identification item
mod <- D2mirt(x, modid="W7Q10")

# Show summary of results
summary(mod)

# Call to D2mirt with three construct vectors indicated with polar angles
# Note that indicating constructs using items is the same as with D3mirt()
mod <- D2mirt(x, modid="W7Q10", con.pol = c(30, 45, 60))

# Show summary of results
summary(mod)

# Call D2mirt() using the orthogonal optional model
# In this example item W7Q16 is removed from the data frame
y <- data.frame(x[,-16])

# Items are constrained to the x and y-axes using
# Nested lists with positive integers as item indicators
# Note that integers indicate where the items are located in the data frame
mod <- D2mirt(y, modid = list(c(1:14),
                            c(15:19)))

# Show summary of results
summary(mod)
}
}
\references{
Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.\cr
https://doi.org/10.18637/jss.v048.i06

Reckase, M. D. (2009). \emph{Multidimensional Item Response Theory}. Springer.

Reckase, M. D. (1985). The Difficulty of Test Items That Measure More Than One Ability. \emph{Applied Psychological Measurement, 9}(4), 401-412.\cr
https://doi.org/10.1177/014662168500900409

Reckase, M. D., & McKinley, R. L. (1991). The Discriminating Power of Items That Measure More Than One Dimension. \emph{Applied Psychological Measurement, 15}(4), 361-373.\cr
https://doi.org/10.1177/014662169101500407
}
\author{
Erik Forsberg
}
