% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3mirt.R
\name{D3mirt}
\alias{D3mirt}
\title{D3mirt Estimation}
\usage{
D3mirt(x, constructs = NULL)
}
\arguments{
\item{x}{Data frame with rows for items and columns for model parameters. The number of columns must be ≥ 4, i.e., three columns for \emph{a} parameters and at least one column for \emph{d} parameters.}

\item{constructs}{Optional. Nested lists with integers indicating construct. Default is \code{constructs = NULL}.}
}
\value{
S3 object with lists of \emph{a} and \emph{d} parameters, multidimensional discrimination (MDISC), multidimensional item difficulty (MDIFF), direction cosines and degrees for vector angles, construct lists, and vector coordinates.
}
\description{
Descriptive multidimensional item response theory modeling (DMIRT; Reckase, 2009, 1985, Reckase & McKinley, 1991), for dichotomous and polytomous items restricted to follow the graded response model (Samejima, 1969) in three dimensions.
}
\details{
The \code{dmirt()} function takes in a data frame of factor slopes (\emph{a}) and difficulty parameters (\emph{d}) from a three-dimensional graded response model fitted with \link[mirt:mirt]{mirt::mirt} (Chalmers, 2012).
The function returns an S3 object containing dmirt estimates that can be graphically displayed with \code{\link[=plot]{plot()}}.
Model parameters for the multidimensional graded response model must be estimated prior using \code{dmirt()} (see examples below) and follow the necessary model specification and limitations.
This includes that items must be set to load on all factors in the multidimensional graded response model and that the number of factors must be three.
}
\section{Within Multidimensional Modelling}{
In DMIRT it is assumed that items in a multidimensional trait space can measure single or multiple latent abilities (Reckase, 2009, 1985, Reckase & McKinley, 1991).
Under this assumption the two parameter graded response model can be used to extract two parameter multidimensional item characteristics.
These include a single multidimensional discrimination (MDISC) parameter and a multidimensional difficulty (MDIFF) index for each item.
The MDISC  \eqn{A_i} for item \eqn{i} represents the highest level of discrimination the item \eqn{i} can achieve located in a multidimensional theta space.

\deqn{MDISC=A_i=\sqrt{\sum_{k=1}^m a^2_{ik}}}{%
MDISC=A_i=\sqrt{\sum_{k=1}^m a^2_{ik}}}

Just as in unidimensional modelling, the  \eqn{A_i} indicates the direction, as seen from the origin of the model, to the point of maximun slope of the item response surface.
The slope is similarly to the unidimensional case assessed as \eqn{A_i/4}.

The item angle orientation is set by taking the direction cosines of \eqn{a_{il}}, i.e., the cosine angle of item \eqn{i} with respect to each trait dimension \eqn{l}.

\deqn{a_{il}= cos^{-1}\left(\frac{a_{il}}{\sqrt{\sum_{k=1}^m a^2_{ik}}}\right)}{%
a_{il}= cos^{-1}\left(\frac{a_{il}}{\sqrt{\sum_{k=1}^m a^2_{ik}}}\right)}

The resulting direction vector is a characteristic of the item that describe the angle orientation of an item in a multidimensional theta space.

\deqn{MDIFF=B_i= \frac{-d_i}{\sqrt{\sum_{k=1}^m a^2_{ik}}}}{%
\frac{-d_i}{\sqrt{\sum_{k=1}^m a^2_{ik}}}}
}

\section{Constructs}{
The user has the option of including constructs in the estimation, by creating one or more nested lists that indicate what items belong to what construct (see the examples section).
From this, the \code{dmirt()} function calculates new direction cosines for the constructs by adding and normalizing the direction cosines for the items contained in each construct list.
The construct vector arrows can contribute to the analysis by visualizing the average direction for a subset set of items.
}

\section{Scaling of Vector Arrows}{
Regarding plotting, the \code{dmirt()} function returns vector coordinates estimated with and without the MDISC as a scalar for arrow length. If the MDISC is not used, all vector arrows are scaled to unit length.
This can help reduce clutter in the graphical output.
}

\section{Model Identification}{
An important part of descriptive multidimensional item response theory modeling is the model identification that is performed before fitting the three-dimensional graded response model.
An example based on a 10-item set in which the model has already been identified is given below.
For more information on model identification, please read the documentation included under \code{D3mirt::modid()}.
}

\examples{
\dontrun{
# Preparation: Fitting a three-dimensional graded response model with orthogoanl factors
# Example uses a scale with 10 items in total, named I_01...I_10
# In the example below item I_01 and item I_10 have been selected to identify the model
# All items in the set are specified to load on all three factors (F1 to F3)
The START and FIXED commands are used to locate the orthogonal axis in the model
library(mirt)
spec <- ' F1 = 1-10
          F2 = 1-10
          F3 = 1-10

          START=(I_05,a2,0)
          START=(I_05,a3,0)

          START=(I_10,a3,0)

          FIXED=(I_05,a2)
          FIXED=(I_05,a3)

          FIXED=(I_10,a3) '


mod1 <- mirt(x, spec, itemtype = 'graded', SE = TRUE, method = 'QMCEM')

# Assign data frame with factor loadings (columns 1-3)
# and difficulty parameters (columns 4-7) from mod1
d <- data.frame(mirt::coef(mod1, simplify=TRUE)$'items'[,1:7])

# Estimation with dmirt(), including nested lists for two constructs
c <- list(list(1,3,4,6,8), list(2,5,7,9,10))
g <- D3mirt(d, c)
}
}
\references{
Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.

Reckase, M. D. (2009). \emph{Multidimensional Item Response Theory}. Springer.

Reckase, M. D. (1985). The Difficulty of Test Items That Measure More Than One Ability. \emph{Applied Psychological Measurement, 9}(4), 401-412–412. https://doi-org.ezp.sub.su.se/10.1177/014662168500900409

Reckase, M. D., & McKinley, R. L. (1991). The Discriminating Power of Items That Measure More Than One Dimension. \emph{Applied Psychological Measurement, 15}(4), 361-373–373. https://doi-org.ezp.sub.su.se/10.1177/014662169101500407

Samejima, F. (1969). Estimation of latent ability using a response pattern of graded scores. \emph{Psychometrika 34}, 1–97. https://doi.org/10.1007/BF03372160
}
\author{
Erik Forsberg
}
