% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{information}
\alias{information}
\title{Item Information and Standard Errors}
\usage{
information(x, ci = 0.95, digits = 4, ...)
}
\arguments{
\item{x}{An x of class \code{D3mirt} or \code{D2mirt}.}

\item{ci}{Critical value for the confidence interval based on percentage. The default is \code{ci = 0.95}.}

\item{digits}{The number of digits shown per estimate. The default is \code{digits = 4}.}

\item{...}{Additional arguments.}
}
\value{
A brief model description followed by descriptives for MDISC and DDISC constructs, such as total sum of information, and the arithmetic mean of item information and standard error.
Tables containing MDISC, DDISC, and standard errors for all items.
}
\description{
A information and standard error analysis function for \code{\link[=D3mirt]{D3mirt()}} and \code{\link[=D2mirt]{D2mirt()}}.
}
\details{
The \code{information()} function extracts item information and standard errors from S3 objects returned by \code{D3mirt()} or \code{D2mirt()}.
The number of decimal places in the output is controlled by the \code{digits} argument.

Standard item information, denoted \eqn{I_i} for item \eqn{i}, is defined as

\deqn{I_i(\theta) := P_i(\theta) Q_i(\theta)a_i^2,}

where \eqn{P_i(\theta)} is the probability of a correct response and \eqn{Q_i(\theta)} the complement on ability \eqn{\theta}.
Taking measurement direction in account gives

\deqn{I_i(\theta, \omega) := P_i(\theta) Q_i(\theta) \bigg ( \sum^n a_i \cos \omega \bigg )^2,}

in which the summation is the DDISC index on \eqn{n} dimensions and measurement angle \eqn{\omega}.
The standard error represents the inverse of the information function and is defined as

\deqn{SE(\hat{a}_i) := \frac{1}{\sqrt{I(\hat{a}_i)}},}

in which \eqn{\hat{a}_i} is the discrimination estimator.
The MDISC and DDISC parameters are presented with an confidence interval based on the standard error, defined as

\deqn{CI = \hat{a}_i \pm z \times SE(\hat{a}_i),}

for critical value \eqn{z}, set by user.

Note that MDISC estimates represent the maximum discrimination of an item at its optimal angle in the multidimensional space.
Because this orientation differs for every item, MDISC values are not directly comparable.
Consequently, summarizing MDISC estimates using measures of central tendency (such as the mean or median) can yield misleading interpretations if assumed to represent a sum of unidimensional information.
In contrast, DDISC estimates are projected onto a specific, fixed construct vector.
Because they are constrained to a single shared orientation, DDISC estimates—and their associated information values—can be meaningfully compared and aggregated under the assumption of unidimensionality.
It should also be mentioned that the calculation of the standard error is capped to avoid extreme value problems (see documentation in \code{\link[=D3mirt]{D3mirt()}} and \code{\link[=D2mirt]{D2mirt()}}).
}
\examples{
\dontrun{
# mod is an S3 x from D3mirt or D2mirt
information(mod)
}
}
\author{
Erik Forsberg
}
