% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dmirt}
\alias{plot.dmirt}
\title{plot.dmirt: Plot function for the D3mirt package}
\usage{
\method{plot}{dmirt}(
  x,
  constructs = FALSE,
  vec.scalar = 1,
  con.scalars = c(1, 1),
  profiles = NULL,
  hide = FALSE,
  items = NULL,
  item.lab = TRUE,
  item.names = NULL,
  construct.lab = NULL,
  adjustlab = c(0.5, -0.8),
  diff.level = NULL,
  background = "white",
  width.rgl.x = 1040,
  width.rgl.y = 1040,
  view = c(15, 20, 0.7),
  axis.scalar = 1.2,
  axis.col = "black",
  axis.points = "black",
  points = TRUE,
  axis.ticks = TRUE,
  nticks = 8,
  title = "",
  line = -5,
  x.lab = "X",
  y.lab = "Y",
  z.lab = "Z",
  show.plane = TRUE,
  plane.color = "grey80",
  type = "rotation",
  col = c("black", "grey20", "grey40", "grey60", "grey80"),
  arrow.width = 0.6,
  n = 20,
  theta = 0.2,
  barblen = 0.03,
  c.type = "rotation",
  c.col = "black",
  c.arrow.width = 0.6,
  c.n = 20,
  c.theta = 0.2,
  c.barblen = 0.03,
  spheres.r = 0.05,
  sphere.col = "grey20",
  ellips = TRUE,
  CI.level = 0.95,
  ellips.col = "grey80",
  ellips.alpha = 0.2,
  ...
)
}
\arguments{
\item{x}{S3 dmirt object}

\item{constructs}{Logical, if construct vector arrows should be plotted. Default set to FALSE}

\item{vec.scalar}{Scalar for changing item arrow length. Default is \code{vec.scalar = 1}.}

\item{con.scalars}{Scalars, for values <0 and >0, for changing construct arrow length. Default is \code{con.scalars = c(1,1)}.}

\item{profiles}{Data frame with coordinates for spheres representing respondents}

\item{hide}{Logical, if items should be plotted. Default is \code{hide = FALSE}.}

\item{items}{Optional. The user can input a list of integers indicating what item vector arrows will be visible while remaining items are only hidden.}

\item{item.lab}{Logical, if item labels should be plotted. Default is \code{item.names = TRUE}.}

\item{item.names}{Optional. The user can provide a list of item names that will override row names extracted from the data frame. Names can be numbers or characters.}

\item{construct.lab}{Optional. User can provide list of names for constructs.}

\item{adjustlab}{Vector of adjustment parameters for position of item and construct labels for the \code{text3d} function. The first value is for horizontal adjustment and the second is for vertical adjustment. Default is \code{adjustlab = c(0.5, -0.8)}.}

\item{diff.level}{Optional. If multiple levels of difficulty are used in the model, it is possible to plot a single level of difficulty by imputing its integer value.}

\item{background}{Set background color for the graphical device, Default is \code{background = "white"}.}

\item{width.rgl.x}{Width in the x direction for \code{par3d()}. Default is \code{width.rgl.x = 1040}.}

\item{width.rgl.y}{Width in the y direction for \code{par3d()}. Default is \code{width.rgl.y = 1040}.}

\item{view}{Vector with polar coordinates and zoom factor for the \code{view3d} function. Default is \code{view = c(15,20, 0.7)}.}

\item{axis.scalar}{Scalar factor the length of all three axis in the 3D model. Default is \code{axis.fac = 1.2}.}

\item{axis.col}{Color of axis for the \code{segment3D()}function, default is \code{axis.col = "Black"}.}

\item{axis.points}{Color of axis points for the \code{points3d()} function. Default is \code{axis.points = "black"}.}

\item{points}{Logical, if axis from the \code{points3d()} should have end points. Default is \code{points = TRUE}.}

\item{axis.ticks}{Logical, if axis ticks from the \code{axis3d()}function should be plotted. Default is `axis.ticks = TRUE'.}

\item{nticks}{Number of ticks for the \code{axis3d()}function. Default is \code{nticks = 8}.}

\item{title}{The main title for the plot plotted with the \code{title3d()} function.}

\item{line}{Title placement for \code{title3d()}. Default is \code{line = -5}.}

\item{x.lab}{Labels for x-axis, Default is \code{x.lab = "X"}.}

\item{y.lab}{Labels for y-axis, Default is \code{y.lab = "Y"}.}

\item{z.lab}{Labels for y-axis, Default is \code{z.lab = "Z"}.}

\item{show.plane}{Logical, if xz-plane should be plotted in graphical device. Default is \code{show.plane = TRUE}.}

\item{plane.color}{Color of plane, default is \code{plane.color = "grey80"}.}

\item{type}{Type of vector arrow for items, default is \code{type = "rotation"}. See \link[rgl:arrow3d]{rgl::arrow3d} for more options regarding arrow types.}

\item{col}{Vector of colors representing difficulty levels for items using the \code{arrow3d()} function. Default is \code{col = c("black", "grey20", "grey40", "grey60", "grey80")}}

\item{arrow.width}{Width of vector arrows for \code{arrow3d()}. Default is \code{arrow.width = 0.6}.}

\item{n}{Number of barbs for the vector arrows from the \code{arrow3d()} function. Default is \code{n = 20}.}

\item{theta}{Opening angle of barbs for vector arrows from \code{arrow3d()}. Default is \code{theta = 0.2}.}

\item{barblen}{The length of the barbs for vector arrows from \code{arrow3d()}. Default is \code{barblen = 0.03}.}

\item{c.type}{Type of vector arrow for constructs. See \link[rgl:arrow3d]{rgl::arrow3d} for more options regarding arrow types. Default is \code{c.type = "rotation"}.}

\item{c.col}{Color for construct vector arrows from \code{arrow3d()}, default is \code{c.col = "black"}.}

\item{c.arrow.width}{Width of construct vector arrows for \code{arrow3d()}. Default is \code{c.arrow.width = 0.6}.}

\item{c.n}{Number of barbs for the construct vector arrows from the \code{arrow3d()} function. Default is \code{c.n = 20}.}

\item{c.theta}{Opening angle of barbs for construct vector arrows from \code{arrow3d()}. Default is \code{c.theta = 0.2}.}

\item{c.barblen}{The length of the barbs for construct vector arrows from \code{arrow3d()}. Default is \code{c.barblen = 0.03}.}

\item{spheres.r}{Radius of sphere for \code{spheres3d()}. Default is \code{spheres.r = 0.05}.}

\item{sphere.col}{Color of sphere \code{spheres3d()}. Default is \code{sphere.col = "grey20"}.}

\item{ellips}{Logical, if spheres should include an ellipsoid outlining an confidence region returned from the \code{ellipse3d()} function. Default is \code{ellips = TRUE}.}

\item{CI.level}{Level of confidence for \code{ellipse3d()}, default is \code{CI.level = 0.95}.}

\item{ellips.col}{Color of the ellipse from \code{ellipse3d()}. Default is \code{ellips.col = "grey80"}.}

\item{ellips.alpha}{Opacity for the confidence region from \code{ellipse3d()}. Default is \code{ellips.alpha = 0.20}.}

\item{...}{Any additional arguments to be passed to RGL}
}
\value{
RGL graphical device.
}
\description{
The \code{plot.dmirt()}visualize the S3 \code{dmirt()} object in a 3D theta space. The plot function is based on \link{rgl} package for data visualization. Output consists of a RGL graphical device that can be exported with dedicated functions (see the examples section).
}
\details{
The RGL device has orthogonal standardized axis centered at 0. The function allows plotting of all items, a selection of items as well as plotting a single item and adding constructs to the graphical output (see examples section). The user can also choose to plot one level of difficulty at a time if multiple levels of difficulty are used in the model. Item names are plotted by default but the user has the option of imputing new names for the items and include names for the constructs.

Moreover, the function can also plots respondents on top of the item model represented as spheres located with the help of respondents factors scores providing the necessary coordinates (see \link{profile} for more details). This allows for profile analysis in which the user can select respondent rows based on some criteria. When analyzing respondent profiles the user has the option of hiding the item vectors to avoid visual cluttering.
More specifically, item location shows the items three dimensional level of difficulty. If Likert items are used, each item will have multiple level of
difficulty and can therefore be said to show the difficulty range of an item. Moreover, the angle of the vector arrows shows the direction of optimal discrimination in the model. In turn, this indicates what traits, one up to three, the item can be said to describe. Lastly, the length of each vector arrows shows the strength of discrimination, in which longer arrows indicate higher discrimination. Short vector arrows are therefore signs of model violations.
}
\examples{
\dontrun{
# Preparation: Calculate dmirt estimates with constructs
c <- list(list(1,3,4,6,8), list(2,5,7,9,10))
g <- dmirt(d, c)


# Plot RGL device with constructs visible
plot.dmirt(g, constructs = TRUE)


# Plot a selection of items from the model
plot.dmirt(g, constructs = TRUE, items = c(1,3,4,6,8))



# Export RGL device in consol
plot.dmirt(g, constructs = TRUE)
rglwidget(width = 1040, height = 1040)


# Export RGL device to file
plot.dmirt(g, constructs = TRUE) # ta bort dmirt?
rgl.snapshot('RGLdevice.png', fmt = 'png')
}
}
