% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.D3mirt.R
\name{print.D3mirt}
\alias{print.D3mirt}
\title{Print Function for S3 Objects of Class \code{D3mirt}}
\usage{
\method{print}{D3mirt}(x, ...)
}
\arguments{
\item{x}{S3 object of class \code{D3mirt}.}

\item{...}{Additional arguments.}
}
\value{
Model summary containing class, number of items, levels of difficulty, the number of construct vectors and their respective items.
}
\description{
The \code{print.D3mirt()} function returns a compressed output from the \code{\link[=D3mirt]{D3mirt()}} function.
}
\examples{
\donttest{
# Load data
data("anes0809offwaves")
x <- anes0809offwaves
x <- x[,3:22] # Remove columns for age and gender

# Fit a three-dimensional graded response model with orthogonal factors
spec <- '  F1 = 1-20
           F2 = 1-20
           F3 = 1-20

           START=(W7Q3,a2,0)
           START=(W7Q3,a3,0)

           START=(W7Q20,a3,0)

           FIXED=(W7Q3,a2)
           FIXED=(W7Q3,a3)

           FIXED=(W7Q20,a3) '


mod.1 <- mirt::mirt(x,
                   spec,
                   itemtype = 'graded',
                   SE = TRUE,
                   method = 'QMCEM')

# Optional: Load the mod.1 data for this example directly from the package file
# load(system.file("mod.1.Rdata", package = "D3mirt"))

# Call D3mirt() and create list of constructs
c <- list(list(1,2,3,4,5,6,7,8,9,10),
          list(11,12,13,14),
          list(15,17,18,19,20))
g <- D3mirt(mod.1, c)

# Print model summary
print(g)
}
}
\author{
Erik Forsberg
}
