% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile}
\alias{profile}
\title{D3mirt Profile Analysis}
\usage{
profile(
  x,
  y = NULL,
  column = 1,
  condition = c(">"),
  prob = 0.75,
  method = "EAP",
  QMC = TRUE
)
}
\arguments{
\item{x}{A three-dimensional S4 \link[mirt:mirt]{mirt::mirt} object, fitted using descriptive multidimensional item response theory modeling (Reckase, 2009)}

\item{y}{A S4 \link[mirt:mirt]{mirt::mirt} object of the same type as above, or a data frame with respondent's scores.}

\item{column}{Numeric indicating what column to use regarding the input for y. Default is \code{column = 1}.}

\item{condition}{String indicating what logical condition to use for filtering. Options are: \code{">"}, \code{"<"}, \code{">="}, and \code{"<="}. Default is \verb{condition = >}.}

\item{prob}{Decimal indicating what probability to use as cut-off. Default is \code{prob = .75}.}

\item{method}{The method used to extract respondent factor scores with \code{\link[mirt:fscores]{mirt::fscores()}}. Default is \code{method = EAP}.}

\item{QMC}{Logical, if Quasi-Monte Carlo integration should be used for extracting respondents' factor scores from \code{\link[mirt:fscores]{mirt::fscores()}}. Default is \code{QMC = TRUE}.}
}
\value{
List of factor scores
}
\description{
\code{profile()} is a wrapper function to \code{\link[mirt:fscores]{mirt::fscores()}} (Chalmers, 2012) and use \code{\link[stats:quantile]{stats::quantile()}} to select respondent by row.
}
\details{
The returned object can be used to plot respondents' scores with \code{\link[=plot]{plot()}}, using factor scores as coordinates in the graphical output. Please consult the documentation on \code{\link[mirt:fscores]{mirt::fscores()}} for more details and available options regarding the \code{fscores()} function.
}
\examples{
\dontrun{
# mod1 is the fitted three dimensional S4 mirt object
# y is a data frame containing scores from the same respondents
# The function call states that rows containing scores higher than .75, of all scores in column 20 from data frame y, will be selected
p <- profile(mod1, y, column = 20, condition = ">", prob = .75)


# Plot model with the D3mirt plot function while hiding item vectors
plot(x, profiles = p, hide = TRUE)
}
}
\references{
Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.

Reckase, M. D. (2009). \emph{Multidimensional Item Response Theory}. Springer.
}
\author{
Erik Forsberg
}
