% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryD2mirt.R
\name{summary.D2mirt}
\alias{summary.D2mirt}
\title{Summary Method for S3 Objects of Class \code{D2mirt}}
\usage{
\method{summary}{D2mirt}(object, digits = 4, ...)
}
\arguments{
\item{object}{A S3 object of class \code{D2mirt}.}

\item{digits}{The number of digits shown per estimate. The default is \code{digits = 4}.}

\item{...}{Additional arguments.}
}
\value{
A brief model description followed by tables containing \emph{a} and \emph{d} parameters, multidimensional discrimination (MDISC), multidimensional item difficulty (MDIFF), direction cosines, and degrees for vector angles for items.
If constructs were used in the estimation process, the summary function will also show tables for direction cosines, degrees for construct vectors, and directional discrimination (DDISC) parameters.
}
\description{
The summary method for the \code{\link[=D2mirt]{D2mirt()}} function.
}
\examples{
\dontrun{
# Load data
data("anes0809offwaves")
x <- anes0809offwaves
x <- x[, 3:22] # Remove columns for age and gender

# Call D2mirt()
mod <- D2mirt(x, modid = "W7Q10")

# Call to summary
summary(mod)

#' # Call to summary rounded off to 2 digits
summary(mod, digits = 2)
}
}
\author{
Erik Forsberg
}
