% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.D3dmirt}
\alias{summary.D3dmirt}
\title{summary.dmirt}
\usage{
\method{summary}{D3dmirt}(object, ..., digits = 4)
}
\arguments{
\item{object}{S3 object of class \code{D3mirt}.}

\item{...}{Additional arguments to be passed to methods.}

\item{digits}{User can adjust number of digiots shown per estimate. Default is \code{digits = 4}.}
}
\value{
Lists # double check
}
\description{
The \code{summary.dmirt()} function presents a compressed output from the \code{dmirt()} S3 object.
}
\details{
The output presents one frame with item factor loadings, difficulty parameters the MDISC estimate for each item. This is followed by a frame with columns for the direction cosines and columns for the direction cosines transformed to degrees. If the S3 object contain constructs, an extra frame containing direction cosines and angles in degrees for the constructs will be included in the output.
}
\examples{
\dontrun{
# Preparation
#' c <- list(list(1,3,4,6,8), list(2,5,7,9,10))
g <- dmirt(d, c)

# Call summary
summary(g)
}
}
