% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.D3mirt.R
\name{summary.D3mirt}
\alias{summary.D3mirt}
\title{Summary Function for S3 Objects of Class \code{D3mirt}}
\usage{
\method{summary}{D3mirt}(object, ..., digits = 4)
}
\arguments{
\item{object}{S3 object of class \code{D3mirt}.}

\item{...}{Additional arguments.}

\item{digits}{The number of digits shown per estimate. The default is \code{digits = 4}.}
}
\value{
Tables containing \emph{a} and \emph{d} parameters, multidimensional discrimination (MDISC), multidimensional item difficulty (MDIFF), direction cosines, and degrees for vector angles for items.
If constructs were used in the estimation process, the summary function will also show tables for direction cosines, and degrees for construct vectors as well as directional discrimination (DDISC) parameters.
}
\description{
The \code{summary.D3mirt()} function returns a compressed output from the \code{\link[=D3mirt]{D3mirt()}} function.
}
\examples{
\donttest{
# Load data
data("anes0809offwaves")
x <- anes0809offwaves
x <- x[,3:22] # Remove columns for age and gender

# Fit a three-dimensional graded response model with orthogonal factors
spec <- '  F1 = 1-20
           F2 = 1-20
           F3 = 1-20

           START=(W7Q3,a2,0)
           START=(W7Q3,a3,0)

           START=(W7Q20,a3,0)

           FIXED=(W7Q3,a2)
           FIXED=(W7Q3,a3)

           FIXED=(W7Q20,a3) '


mod1 <- mirt::mirt(x,
                   spec,
                   itemtype = 'graded',
                   SE = TRUE,
                   method = 'QMCEM')

# Optional: Load the mod1 data for this example directly from the package file
# load(system.file("mod1.Rdata", package = "D3mirt"))

# Call D3mirt() and create list of constructs
c <- list(list(1,2,3,4,5,6,7,8,9,10),
          list(11,12,13,14),
          list(15,17,18,19,20))
g <- D3mirt(mod1, c)

# Call to summary
summary(g)
}
}
\author{
Erik Forsberg
}
