% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dmirt}
\alias{summary.dmirt}
\title{D3mirt Summary Function}
\usage{
\method{summary}{dmirt}(object, ..., digits = 4)
}
\arguments{
\item{object}{S3 object of class \code{dmirt}.}

\item{...}{Additional arguments to be passed to methods.}

\item{digits}{User can adjust the number of digits shown per estimate. Default is \code{digits = 4}.}
}
\value{
Lists containing \emph{a} and \emph{d} parameters, multidimensional discrimination (MDISC), multidimensional item difficulty (MDIFF), direction cosines, and degrees for vector angles for items and constructs.
}
\description{
The \code{summary.dmirt()} function presents a compressed output from the \code{\link[=D3mirt:dmirt]{D3mirt:dmirt()}} S3 object.
}
\examples{
\dontrun{
# Preparation
#' c <- list(list(1,3,4,6,8), list(2,5,7,9,10))
g <- dmirt(d, c)

# Call summary
summary(g)
}
}
\author{
Erik Forsberg
}
