% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.modid.R
\name{summary.modid}
\alias{summary.modid}
\title{Summary Function for S3 Objects of Class \code{modid}}
\usage{
\method{summary}{modid}(object, ..., digits = 4)
}
\arguments{
\item{object}{S3 object of class \code{modid}.}

\item{...}{Additional arguments.}

\item{digits}{The number of digits shown per estimate. The default is \code{digits = 4}.}
}
\value{
Model identification items (one less than the number of factors), factor loadings and absolute sum score for items, and squared factor loadings for the model identification items.
}
\description{
The \code{summary.modid()} function returns a compressed output from the \code{\link[=modid]{modid()}} function.
}
\examples{
\donttest{
# Load data
data("anes0809offwaves")
x <- anes0809offwaves
x <- x[,3:22] # Remove columns for age and gender

# Identify the DMIRT model using a three-factor EFA
g <- modid(x)

# Optional: Load the EFA data for this example directly from the package file
# load(system.file("efa.Rdata", package = "D3mirt"))

# Call to summary
summary(g)
}
}
\author{
Erik Forsberg
}
