% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3mirt.R
\name{D3mirt}
\alias{D3mirt}
\title{3D DMIRT Modeling}
\usage{
D3mirt(x, constructs = NULL)
}
\arguments{
\item{x}{Data frame with rows for items and columns for model parameters. The number of columns must be more than or equal to 4, i.e., three columns for slope (\emph{a}) parameters and at least one column for difficulty (\emph{d}) parameters.}

\item{constructs}{Optional. Nested lists with integer values indicating construct. Default is \code{constructs = NULL}.}
}
\value{
S3 object of class \code{D3mirt} with lists of \emph{a} and \emph{d} parameters, multidimensional discrimination (MDISC), multidimensional item difficulty (MDIFF), direction cosines and degrees for vector angles, construct lists, and vector coordinates.
}
\description{
Descriptive multidimensional item response theory modeling (DMIRT; Reckase, 2009, 1985, Reckase & McKinley, 1991), for dichotomous and polytomous items in three dimensions.
}
\details{
The \code{D3mirt()} function takes in a data frame of factor slopes and difficulty parameters from a three-dimensional graded
response model, fitted in accordance with descriptive item response theory model specifications described below,
and returns an S3 object containing estimates that can be graphed as vector arrows with \link{plotD3mirt}.
}
\section{Model Identification}{
Model parameters from the multidimensional graded response model must be assessed prior to using the \code{D3mirt()} function.
This implies that all items are set to load on all factors in the model by choosing and fixing two items from the scale or set (see examples section below).
The \code{modid()} function can help with the latter by suggesting what items to use.
For more information on model identification, see documentation regarding \link{modid}.
}

\section{Within Multidimensional Modeling}{
In DMIRT analysis, also called within multidimensional modeling, it is assumed that items in a multidimensional trait space
can measure single or multiple latent abilities (Reckase, 2009, 1985, Reckase & McKinley, 1991).
The DMIRT method is said to be descriptive because the estimates describes item characteristics when more than one latent dimension is used in the analysis.
Under the assumption of within-dimensionality, the two-parameter graded response model is used to extract two-parameter multidimensional item characteristics.
These include a single multidimensional discrimination (MDISC) parameter and a multidimensional difficulty (MDIFF) index for each item.

The MDISC  \eqn{A_i} for item \eqn{i} represents the highest level of discrimination the item \eqn{i} can achieve located
in a multidimensional theta space, with \eqn{m} number of dimensions and \eqn{a_{ik}} item slope parameters.

\deqn{MDISC=A_i=\sqrt{\sum_{k=1}^m a^2_{ik}}}{%
MDISC=A_i=\sqrt(\sum_k=1^m a^2_ik)}

Just as in unidimensional modeling, the  \eqn{A_i} indicates the direction, as seen from the origin of the model, to the point of maximum slope of the item response surface.
The slope is, similarly to the unidimensional case, assessed as \eqn{A_i/4} (omitted in the equation above).

The items angle orientation is set by taking the direction cosines, in linear algebra terms, of \eqn{a_{il}}, i.e., the slope values of item \eqn{i} on coordinate axis \eqn{l}.

\deqn{a_{il}= cos^{-1}\left(\frac{a_{il}}{\sqrt{\sum_{k=1}^m a^2_{ik}}}\right)}{%
a_il=cos^-1(a_il/(\sqrt(\sum_k=1^m a^2_ik))}

The resulting direction vector is a characteristic of the item that describes the angular orientation of an item in a multidimensional theta space.

The multidimensional version of the difficulty parameter, \eqn{B_i}, for item \eqn{i} is defined as the negative intercept \eqn{d_i} divided by the MDISC.

\deqn{MDIFF=B_i=\frac{-d_i}{\sqrt{\sum_{k=1}^m a^2_{ik}}}}{%
-d_i/(\sqrt(\sum_k=1^m a^2_ik))}

The MDIFF is interpreted similarly as the difficulty parameter in the unidimensional model.
That is, higher values indicate that higher levels of ability for a probability of a correct response of more than .5 are necessary.
Moreover, the MDIFF, just as in the unidimensional model, sets the distance along the x-axis from the origin of the model to the point of maximum slope.
However, in DMIRT analysis, the MDIFF functions as a multidimensional location parameter that indicate the distance from the origin to the point of maximum slope
following the direction by the \eqn{a_{il}} equation.

The length of the vector arrows is set by taking the MDIFF times the MDISC so that items with higher discrimination have longer vector arrows.
}

\section{Limitations}{
The \code{D3mirt} technique is based on the grade response item model (Samejima, 1969) extended to a multidimensional space.
Consequently, the latter entails that \code{D3mirt} analysis is limited to items that fit the GRM.
Moreover, for the \code{D3mirt}analysis the number of dimensions can be up to three.
However, due to the nature of the model specification, the analysis can handle dimensions less than three.
This since the third axis, the z-axis, is free while only two items must meet the model identification requirements to locate the first (x-axis) and second axis (y-axis).
}

\section{User Options}{
\subsection{Constructs}{

The user has the option of including constructs in the estimation, by creating one or more nested lists that indicate what items belong to what construct (see the examples section).
From this, the \code{D3mirt()} function calculates direction cosines for the constructs by adding and normalizing the direction cosines for the items contained in each construct list.
The construct vector arrows can contribute to the analysis by visualizing the average direction for a subset set of items.
}

\subsection{Scaling of Item Vector Arrows}{

Regarding plotting, the \code{D3mirt()} function returns vector coordinates estimated with and without the MDISC as a scalar for arrow length.
If the object is plotted without the MDISC , all vector arrows are scaled to one unit length.
This can help reduce clutter in the graphical output when using \code{plotD3mirt()}.
}
}

\examples{
# Load data
data("anes08_09offwaves")
x <- anes08_09offwaves
x <- x[,3:22] # Remove columns for age and gender

# Fit a three-dimensional graded response model with orthogonal factors
# Example below use Likert items from the built in data set "anes08_09offwaves"
# Item W7Q3 and item W7Q3 have been selected with `modid()`
# The model specification below specify all items (1-20) to load on all three factors (F1-F3)
# The START and FIXED commands are used to identify the orthogonal structure in the model
spec <- ' F1 = 1-20
          F2 = 1-20
          F3 = 1-20

          START=(W7Q3,a2,0)
          START=(W7Q3,a3,0)

          START=(W7Q20,a3,0)

          FIXED=(W7Q3,a2)
          FIXED=(W7Q3,a3)

          FIXED=(W7Q20,a3) '


mod1 <- mirt::mirt(x, spec, itemtype = 'graded', SE = TRUE, method = 'QMCEM')

# Assign data frame with factor loadings (located in the first three columns in mod1),
# and difficulty parameters (columns 4-7 in mod1) with mirt::coef and $'items'[,1:7]))
d <- data.frame(mirt::coef(mod1, simplify=TRUE)$'items'[,1:7])

# Call to `D3mirt()`, including optional nested lists for two constructs
# Item W7Q16 is not included in any construct because of model violations
# The model violations will be possible to see when graphing the model
c <- list(list (1,2,3,4), list(5,7,8,9,10), list(11,12,13,14,15,15,16,17,18,19,20))
g <- D3mirt(d, c)
}
\references{
Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.

Reckase, M. D. (2009). \emph{Multidimensional Item Response Theory}. Springer.

Reckase, M. D. (1985). The Difficulty of Test Items That Measure More Than One Ability. \emph{Applied Psychological Measurement, 9}(4), 401-412–412. https://doi-org.ezp.sub.su.se/10.1177/014662168500900409

Reckase, M. D., & McKinley, R. L. (1991). The Discriminating Power of Items That Measure More Than One Dimension. \emph{Applied Psychological Measurement, 15}(4), 361-373–373. https://doi-org.ezp.sub.su.se/10.1177/014662169101500407

Samejima, F. (1969). Estimation of latent ability using a response pattern of graded scores. \emph{Psychometrika 34}, 1–97. https://doi.org/10.1007/BF03372160
}
\author{
Erik Forsberg
}
