% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmirt.R
\name{dmirt}
\alias{dmirt}
\title{D3mirt S3 Object}
\usage{
dmirt(x, constructs = NULL)
}
\arguments{
\item{x}{Data frame with rows for items and columns for model parameters. The number of columns must be ≥ 4, i.e., three columns for \emph{a} parameters and at least one column for \emph{d} parameters.}

\item{constructs}{Optional. Nested lists with integers indicating construct. Default is \code{constructs = NULL}.}
}
\value{
S3 object with lists of \emph{a} and \emph{d} parameters, multidimensional discrimination (MDISC), multidimensional item difficulty (MDIFF), direction cosines and degrees for vector angles, construct lists, and vector coordinates.
}
\description{
Descriptive multidimensional item response theroy modeling, following Reckase (2009), restricted to three dimensions.
The \code{dmirt()} function takes in a data frame of factor slopes (\emph{a}) and difficulty parameters (\emph{d}) from a three-dimensional graded response model,
fitted with \link[mirt:mirt]{mirt::mirt} (Chalmers, 2012). The function returns an S3
object containing dmirt estimates that can be graphically displayed with \code{\link[=plot]{plot()}}.
}
\details{
Model parameters for the multidimensional graded response model must be estimated before using \code{dmirt()} (see examples below) and follow the necessary model specification and limitations.
This includes that items must be set to load on all factors in the graded response model and that the number of factors must be three.

The user has the option of including constructs in the estimation, by creating one or more nested lists that indicate what items belong to what construct (see the examples section).
From this, the \code{dmirt()} function calculates construct direction cosines by adding and normalizing the direction cosines for the items contained in each construct list.
In so doing, the constructs vector arrows can contribute to the analysis by visualizing the average direction for a subset set of items.

Regarding plotting, the \code{dmirt()} function returns vector coordinates estimated with and without the MDISC as a scalar for arrow length. If the MDISC is not used, all vector arrows are scaled to one.
This can be useful to reduce clutter in the graphical output.

An important part of the dmirt process is the model identification that is performed before fitting the three-dimensional graded response model.
An example based on a 10-item set in which the model has already been identified is given below.
For more information on model identification please study the documentation included under \code{D3mirt::modid()}.
}
\examples{
\dontrun{
# Preparation: Fitting a three-dimensional graded response model
# Example uses a scale with 10 items in total, named I_01...I_10
# In the example below item I_01 and item I_10 have been selected to identify the model
# All items in the set are specified to load on all three factors (F1 to F3)
The START and FIXED commands are used to locate the orthogonal axis in the model
library(mirt)
spec <- ' F1 = 1-10
          F2 = 1-10
          F3 = 1-10

          START=(I_05,a2,0)
          START=(I_05,a3,0)

          START=(I_10,a3,0)

          FIXED=(I_05,a2)
          FIXED=(I_05,a3)

          FIXED=(I_10,a3) '


mod1 <- mirt(x, spec, itemtype = 'graded', SE = TRUE, method = 'QMCEM')

# Assign data frame with factor loadings (columns 1-3)
# and difficulty parameters (columns 4-7) from mod1
d <- data.frame(mirt::coef(mod1, simplify=TRUE)$'items'[,1:7])

# Estimation with dmirt(), including nested lists for two constructs
c <- list(list(1,3,4,6,8), list(2,5,7,9,10))
g <- dmirt(d, c)
}
}
\references{
Chalmers, R., P. (2012). mirt: A Multidimensional Item Response Theory Package for the R Environment. \emph{Journal of Statistical Software, 48}(6), 1-29.

Reckase, M. D. (2009). \emph{Multidimensional Item Response Theory}. Springer.
}
\author{
Erik Forsberg
}
