% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modid.R
\name{modid}
\alias{modid}
\title{Model identification}
\usage{
modid(x, head = TRUE)
}
\arguments{
\item{x}{A data frame or matrix, ≥ 4 columns, with factor loadings and difficulty parameters estimated with the graded response model as item model.}

\item{head}{Logical, should output contain first five rows per factor only? Default is TRUE.}
}
\value{
The output consists of lists, one per factor from the fitted factor model, containing one column for the factor loadings from each item on the factor interest, and one column with absolute sum scores calculated from remaining factor loadings in the model. Each list is sorted with the lowest absolute sum score highest, and the user is advised to select two items from two different factors in the output that best meet the assumption of orthogonality, i.e., preferably items with absolute sum scores of < .10 and having factor loading > .80. If these conditions cannot be meet the user is advised to proceed with caution since the loading structure implies that an orthogonal structure may not be empirically attainable.
}
\description{
Model identification
}
\examples{
library(psych)
f <- fa(x, nfactors = 3, rotate = "oblimin", residuals = FALSE, SMC=FALSE)
z <- data.frame(f$loadings[,])
modid(z)
}
