% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printD3mirt.R
\name{print.D3mirt}
\alias{print.D3mirt}
\title{Print Method for S3 Objects of Class \code{D3mirt}}
\usage{
\method{print}{D3mirt}(x, ...)
}
\arguments{
\item{x}{A S3 object of class \code{D3mirt}.}

\item{...}{Additional arguments.}
}
\value{
A printed message reporting the number of items, levels of difficulty, the number of construct vectors and the row names of the respective items contained in each construct.
}
\description{
The print method for the \code{\link[=D3mirt]{D3mirt()}} function.
}
\examples{
\dontrun{
# Call D3mirt()
# The first argument can be data frame with model parameters
# or an S4 object of class 'SingleGroupClass' exported from mirt::mirt
g <- D3mirt(mod1)

# Print model summary
print(g)
}
}
\author{
Erik Forsberg
}
