% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.modid.R
\name{print.modid}
\alias{print.modid}
\title{Print Function for S3 Objects of Class \code{modid}}
\usage{
\method{print}{modid}(x, ...)
}
\arguments{
\item{x}{S3 object of class \code{modid}.}

\item{...}{Additional arguments.}
}
\value{
Printed message reporting the number of factors and the suggested model identification items.
}
\description{
The \code{print.D3mirt()} function returns a compressed output from the \code{\link[=modid]{modid()}} function.
}
\examples{
\donttest{
# Load data
data("anes0809offwaves")
x <- anes0809offwaves
x <- x[,3:22] # Remove columns for age and gender

# Identify the DMIRT model using a three-factor EFA
g <- modid(x)

# Optional: Load the EFA data for this example directly from the package file
# load(system.file("efa.Rdata", package = "D3mirt"))

# Print model summary
print(g)
}
}
