% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile}
\alias{profile}
\title{Profile}
\usage{
profile(
  x,
  y = NULL,
  column = 1,
  condition = c(">"),
  prob = 0.75,
  method = "EAP",
  full.scores = TRUE,
  full.scores.SE = FALSE,
  QMC = TRUE
)
}
\arguments{
\item{x}{S4 mirt object.}

\item{y}{S4 mirt object or data frame with respondents scores.}

\item{column}{Select what column in the input for y to use for selection of respondent rows. Default is 1.}

\item{condition}{String input on what logical condition to use for filtering respondents. Options are: ">", "<", ">=", and "<=" and default is set to ">".}

\item{prob}{Indicate what probability to use as cut off, default is .75.}

\item{method}{The method used to extract respondent factor scores from the S4 mirt object. Default is "EAP".}

\item{full.scores}{Logical, should full scores from the \code{fscores()} function be printed? Default is TRUE.}

\item{full.scores.SE}{Logical, should standard errors from the \code{fscores()} function be printed? Default is FALSE.}

\item{QMC}{Integration method for extracting respondents factor scores from the \code{fscores()} function. Default is "QMC".}
}
\value{
List of factor scores from the \code{mirt} S4 object.
}
\description{
\code{profile()} is a wrapper function to \code{fscores()} from the \code{mirt} package. Respondents factor scores are selected row wise based on the logical condition set by user. Factor scores can then be used to plot respondents as spheres in the 3D \code{dmirt} object. Please consult documentation on \code{\link[mirt:fscores]{mirt::fscores()}} for more details and available options regarding the \code{fscores()} function.
}
\examples{
# x is the fitted three dimensional S4 mirt object
# y is a data frame containing scores from the same respondents
p <- profile(x, y, column = 20, condition = ">", prob = .75)
# To plot with the `D3mirt` plot function while hiding item vectors
plot(x, profiles = p, hide = TRUE)
}
