% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.D3mirt}
\alias{summary.D3mirt}
\title{D3mirt Summary Function}
\usage{
\method{summary}{D3mirt}(object, ..., digits = 4)
}
\arguments{
\item{object}{S3 object of class \code{D3mirt}.}

\item{...}{Additional arguments to be passed to methods.}

\item{digits}{User can adjust the number of digits shown per estimate. Default is \code{digits = 4}.}
}
\value{
Lists containing \emph{a} and \emph{d} parameters, multidimensional discrimination (MDISC), multidimensional item difficulty (MDIFF), direction cosines, and degrees for vector angles for items and constructs.
}
\description{
The \code{summary.D3mirt()} function presents a compressed output from the \code{\link[=D3mirt]{D3mirt()}} S3 object.
}
\examples{
\dontrun{
# Preparation: Call `D3mirt()` and create list of constructs
c <- list(list(1,3,4,6,8), list(2,5,7,9,10))
g <- D3mirt(d, c)

# Call to summary
summary(g)
}
}
\author{
Erik Forsberg
}
